Mobile sensor Cloud as Infrastructure Service
The project contains HTML, CCS, Javascript and PHP code which was hosted
in Rackspace and IBM Bluemix. The project was presented for CMPE 281
Cloud Course . It proposes simulated mobile sensors as an Infrastructure
Service. The User Dashboard provides user to add/remove, start/stop
sensors. Interactive Highcharts and D3 charts provides easy way for
users to understand the data.